# apps/api/src/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "walk / dog API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/owner": {
      "get": {
        "tags": ["owners"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Current Owner" },
          "401": { "description": "Unauthenticated" },
          "500": { "description": "Internal server error" }
        }
      },
      "patch": {
        "tags": ["owners"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Updated Owner" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthenticated" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/v1/dogs": {
      "get": {
        "tags": ["dogs"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Owner dogs" },
          "401": { "description": "Unauthenticated" },
          "500": { "description": "Internal server error" }
        }
      },
      "post": {
        "tags": ["dogs"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "201": { "description": "Created Dog" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthenticated" },
          "409": { "description": "Duplicate dog name" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/v1/dogs/{dogId}": {
      "get": {
        "tags": ["dogs"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "dogId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Dog profile" },
          "401": { "description": "Unauthenticated" },
          "404": { "description": "Not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}